+++ src/index.js
@@ -17,7 +17,6 @@ import {
   SimpleColorPicker,
   CarouselEditBlock,
   CarouselViewBlock,
-  ObjectListWidget,
   SliderEditBlock,
   SliderViewBlock,
   TeaserGridEditBlock,
@@ -30,11 +29,8 @@ import {
   HighlightSliderEditBlock,
   TextPillEditBlock,
   TextPillViewBlock,
-  ObjectByType,
   ButtonViewBlock,
   ButtonEditBlock,
-  QuerystringSidebarWidget,
-  QueryStringSortOnWidget,
 } from '@kitconcept/volto-blocks/components';
 import { ImagesGridIconsVariationSchemaExtender } from '@kitconcept/volto-blocks/components/ImagesGrid/schema';
 
@@ -239,12 +235,7 @@ const customBlocks = {
 //   groupBlocksOrder: customGroupBlocksOrder,
 // };
 export default (config) => {
-  config.widgets.widget.object_list = ObjectListWidget;
   config.widgets.widget.style_simple_color = SimpleColorPicker;
-  config.widgets.widget.object_by_type = ObjectByType;
-  config.widgets.widget.query_sort_on = QueryStringSortOnWidget;
-  config.widgets.widget.query = QuerystringSidebarWidget;
-  config.widgets.id.query = QuerystringSidebarWidget;
 
   if (serverConfig) {
     config.settings.expressMiddleware = [