+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Internal
 
 - Add transltion for TextWithPillStyle @iFlameing
+- Translations for the Text with background color in Volto 14a9 @sneridagh
 
 ## 8.0.0 (2021-07-06)