+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking
 
+- Deprecate forked `ObjectListWidget`, `ObjectWidget`, `objectByType`, and custom pre-released new-style `QuerytringSidebarWidget` and `QueryStringSortOnWidget` in favor of the released in Volto ones. @sneridagh
+
 ### Feature
 
 - Custom schema enhancer for accordion block and CSS enhancements @sneridagh