+++ CHANGELOG.md
@@ -5,12 +5,11 @@
 ### Breaking
 
 - Deprecate forked `ObjectListWidget`, `ObjectWidget`, `objectByType`, and custom pre-released new-style `QuerytringSidebarWidget` and `QueryStringSortOnWidget` in favor of the released in Volto ones. @sneridagh
+- This version is compatible with the new i18n architecture in Volto 14 @sneridagh
 
 ### Feature
 
 - Custom schema enhancer for accordion block and CSS enhancements @sneridagh
-
-### Bugfix
 
 ### Internal